crm-project/backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
drf-spectacular==0.27.2
django-filter==23.3
channels==4.0.0
channels-redis==4.1.0
celery==5.3.6
redis==5.0.1
boto3==1.28.63
django-storages==1.14.2
gunicorn==21.2.0
